// src/api/models/getStoreDetailResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Allta API
 * Allta API 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { StoreDetailItem } from './storeDetailItem';

export interface GetStoreDetailResponse {
  /** 결과 여부 */
  ok: boolean;
  /** 매장 상세 정보 */
  store: StoreDetailItem;
}